Re-work deployer logic to use existing services
The deployer currently always creates a free tier conversation
service and uploads a new workspace to that service. That leads to
many services and being unable to find their newly uploaded
workspace from the devops pipeline (it links to whatever it finds
first in the backend).

The idea here is to create a new service only if one doesn't already
exist. If one exists, use it. If more than 1 exists, then use the
first found, this isn't any worse than the current situation.